clean_dir check if is dir

+++ src/cirrus/core/utils/misc.py
@@ -54,6 +54,10 @@ def relative_to_cwd(path: Path) -> Path:
 
 def clean_dir(directory: Path) -> None:
     import shutil
+
+    if not directory.is_dir():
+        return
+
     for f in directory.iterdir():
         if not f.is_symlink() and f.is_dir():
             shutil.rmtree(f)
